requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
django==4.2.8
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   djangorestframework
    #   drf-yasg
django-cors-headers==4.3.1
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-gis
    #   drf-yasg
djangorestframework-gis==1.0
    # via -r requirements.in
drf-yasg==1.21.7
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-yasg
mysqlclient==2.2.0
    # via -r requirements.in
packaging==23.2
    # via
    #   drf-yasg
    #   gunicorn
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via drf-yasg
sqlparse==0.4.4
    # via django
typing-extensions==4.8.0
    # via asgiref
uritemplate==4.1.1
    # via drf-yasg
